+++ bigmler/bigmler.py
@@ -49,7 +49,22 @@ import sys
 from bigmler.dispatcher import main_dispatcher
 from bigmler.analyze.dispatcher import analyze_dispatcher
 from bigmler.cluster.dispatcher import cluster_dispatcher
+from bigmler.delete.dispatcher import delete_dispatcher
 from bigmler.parser import SUBCOMMANDS
+
+
+def check_delete_option(args):
+    """Checks if the --delete option (to be deprecated) is used and changes
+       its syntax to the corresponding subcommand
+
+    """
+    try:
+        delete_index = args.index('--delete')
+        args[0] = "delete"
+        del args[delete_index]
+    except ValueError:
+        pass
+    return args
 
 
 def main(args=sys.argv[1:]):
@@ -62,12 +77,17 @@ def main(args=sys.argv[1:]):
             new_args.extend(args)
         else:
             new_args = args
+        # checks if the old --delete syntax is used
+        new_args = check_delete_option(new_args)        
         if new_args[0] == "main":
+            new_args = check_delete_option(new_args)
             main_dispatcher(args=new_args)
         elif new_args[0] == "analyze":
             analyze_dispatcher(args=new_args)
         elif new_args[0] == "cluster":
             cluster_dispatcher(args=new_args)
+        elif new_args[0] == "delete":
+            delete_dispatcher(args=new_args)
     else:
         sys.exit("BigMLer used with no arguments. Check:\nbigmler --help\n\nor"
                  "\n\nbigmler analyze --help\n\n for a list of options")